+++ edison-jobs/src/test/java/de/otto/edison/jobs/configuration/JobsConfigurationTest.java
@@ -3,12 +3,10 @@ package de.otto.edison.jobs.configuration;
 import de.otto.edison.jobs.definition.DefaultJobDefinition;
 import de.otto.edison.jobs.definition.JobDefinition;
 import de.otto.edison.jobs.service.JobDefinitionService;
-import de.otto.edison.jobs.service.JobMutexGroups;
 import de.otto.edison.jobs.status.JobStatusCalculator;
 import de.otto.edison.status.domain.StatusDetail;
 import de.otto.edison.status.indicator.CompositeStatusDetailIndicator;
 import de.otto.edison.status.indicator.StatusDetailIndicator;
-import org.junit.Before;
 import org.junit.Test;
 
 import java.util.HashMap;
@@ -18,7 +16,6 @@ import static de.otto.edison.status.domain.Status.OK;
 import static java.time.Duration.ofSeconds;
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptyList;
-import static java.util.Collections.emptySet;
 import static java.util.Collections.singletonList;
 import static java.util.Optional.of;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -31,14 +28,13 @@ import static org.mockito.Mockito.when;
 
 public class JobsConfigurationTest {
 
-    private JobMutexGroups jobMutexGroups = new JobMutexGroups();;
     private JobsProperties jobsProperties = new JobsProperties();;
 
     @Test
     public void shouldIndicateOkIfNoJobDefinitionsAvailable() {
         // given
 
-        final JobsConfiguration jobsConfiguration = new JobsConfiguration(jobsProperties, jobMutexGroups);
+        final JobsConfiguration jobsConfiguration = new JobsConfiguration(jobsProperties);
 
         final JobStatusCalculator defaultCalculator = mock(JobStatusCalculator.class);
         when(defaultCalculator.getKey()).thenReturn("warningOnLastJobFailed");
@@ -59,7 +55,7 @@ public class JobsConfigurationTest {
     @Test
     public void shouldConstructCompositeStatusDetailIndicator() {
         // given
-        final JobsConfiguration jobsConfiguration = new JobsConfiguration(jobsProperties, jobMutexGroups);
+        final JobsConfiguration jobsConfiguration = new JobsConfiguration(jobsProperties);
 
         final JobStatusCalculator defaultCalculator = mock(JobStatusCalculator.class);
         when(defaultCalculator.getKey()).thenReturn("warningOnLastJobFailed");
@@ -76,7 +72,7 @@ public class JobsConfigurationTest {
     @Test
     public void shouldUseDefaultJobStatusDetailIndicator() {
         // given
-        final JobsConfiguration jobsConfiguration = new JobsConfiguration(jobsProperties, jobMutexGroups);
+        final JobsConfiguration jobsConfiguration = new JobsConfiguration(jobsProperties);
 
         final JobStatusCalculator defaultCalculator = mock(JobStatusCalculator.class);
         when(defaultCalculator.getKey()).thenReturn("warningOnLastJobFailed");
@@ -97,7 +93,7 @@ public class JobsConfigurationTest {
             put("test", "errorOnLastJobFailed");
         }});
 
-        final JobsConfiguration jobConfiguration = new JobsConfiguration(jobsProperties, jobMutexGroups);
+        final JobsConfiguration jobConfiguration = new JobsConfiguration(jobsProperties);
 
         final JobStatusCalculator defaultCalculator = mock(JobStatusCalculator.class);
         when(defaultCalculator.getKey()).thenReturn("warningOnLastJobFailed");
@@ -121,7 +117,7 @@ public class JobsConfigurationTest {
             put("soMe-TeSt job", "errorOnLastJobFailed");
         }});
 
-        final JobsConfiguration jobConfiguration = new JobsConfiguration(jobsProperties, jobMutexGroups);
+        final JobsConfiguration jobConfiguration = new JobsConfiguration(jobsProperties);
 
         final JobStatusCalculator defaultCalculator = mock(JobStatusCalculator.class);
         when(defaultCalculator.getKey()).thenReturn("warningOnLastJobFailed");